:star2: [update] entity 설계 (수정)
파일 수정 및 폴더 옮김
