--- src/main/java/com/app/demo/entity/Preference.java
@@ -1,4 +1,0 @@
-package com.app.demo.entity;
-
-public enum Preference {
-}
