+++ src/main/java/com/app/demo/entity/User.java
@@ -1,6 +1,7 @@
 package com.app.demo.entity;
 
 
+import com.app.demo.entity.enums.Preference;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -15,11 +16,11 @@ import java.util.List;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @RequiredArgsConstructor
 @Entity
-@Table(name = "member")
-public class Member {
+@Table(name = "user")
+public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_id")
+    @Column(name = "user_id")
     private Long id;
 
     @Column(nullable = false, unique = true, length = 30)
