+++ src/main/java/com/app/demo/entity/enums/Preference.java
@@ -1,0 +1,4 @@
+package com.app.demo.entity.enums;
+
+public enum Preference {
+}
