+++ src/main/java/com/app/demo/entity/mapping/AIPlaylistMusic.java
@@ -1,5 +1,8 @@
-package com.app.demo.entity;
+package com.app.demo.entity.mapping;
 
+import com.app.demo.entity.AIPlaylist;
+import com.app.demo.entity.User;
+import com.app.demo.entity.Music;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -10,18 +13,21 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Entity
-@Table(name = "preferencePlaylistMusic")
+@Table(name = "aiPlaylistMusic")
 
-public class PreferencePlaylistMusic {
+public class AIPlaylistMusic {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "preferencePlaylistMusic_id")
+    @Column(name = "aiPlaylistMusic_id")
     private Long id;
 
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "aiPlaylist_id")
+    private AIPlaylist aiPlaylist;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "preferencePlaylist_id")
-    private PreferencePlaylist preferencePlaylist;
+    @JoinColumn(name = "user_id")
+    private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "music_id")
