+++ src/main/java/com/app/demo/entity/mapping/PreferencePlaylistMusic.java
@@ -1,5 +1,7 @@
-package com.app.demo.entity;
+package com.app.demo.entity.mapping;
 
+import com.app.demo.entity.Music;
+import com.app.demo.entity.PreferencePlaylist;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -10,21 +12,18 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Entity
-@Table(name = "aiPlaylistMusic")
+@Table(name = "preferencePlaylistMusic")
 
-public class AIPlaylistMusic {
+public class PreferencePlaylistMusic {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "aiPlaylistMusic_id")
+    @Column(name = "preferencePlaylistMusic_id")
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "aiPlaylist_id")
-    private AIPlaylist aiPlaylist;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
+    @JoinColumn(name = "preferencePlaylist_id")
+    private PreferencePlaylist preferencePlaylist;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "music_id")
